# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Equity Management API",
        "description": "API documentation for Equity Management System",
        "version": "1.0.1"
    },
    "paths": {
...
        "/api/client-holdings": {
            "get": {
                "tags": [
                    "ClientHoldings"
                ],
                "summary": "List all client holdings",
                "operationId": "0d88c199205776026ff9943ac54e7e18",
                "responses": {
                    "200": {
                        "description": "List of client holdings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClientHolding"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ClientHoldings"
                ],
                "summary": "Create a new holding for a client",
                "operationId": "f3459410ce655f7cef32bb3502d8a2bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "stock_symbol",
                                    "quantity",
                                    "purchase_price"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "stock_symbol": {
                                        "type": "string",
                                        "example": "DSEXYZ"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "purchase_price": {
                                        "type": "number",
                                        "example": 75.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Holding created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientHolding"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client-holdings/{id}": {
            "get": {
                "tags": [
                    "ClientHoldings"
                ],
                "summary": "Get a specific client holding",
                "operationId": "d6c28f7e8f1aeb3cba25836a473546e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Holding found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientHolding"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ClientHoldings"
                ],
                "summary": "Update a client holding",
                "operationId": "32cbdabf2adef0c474dfe15039e82025",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "stock_symbol": {
                                        "type": "string",
                                        "example": "DSEXYZ"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "purchase_price": {
                                        "type": "number",
                                        "example": 75.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Holding updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientHolding"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ClientHoldings"
                ],
                "summary": "Delete a client holding",
                "operationId": "b5e7ba906e8f79bed57eed479cbccc99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client-holdings/import": {
            "post": {
                "tags": [
                    "ClientHoldings"
                ],
                "summary": "Import client holdings from Excel or CSV",
                "operationId": "d3efa91c8742461d53bc252e2347ffeb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Excel or CSV file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Import successful"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed or file format error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "File format error: Invalid file format"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Import failed due to other errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Import failed: Unexpected error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client-holdings/update-prices": {
            "post": {
                "tags": [
                    "ClientHoldings"
                ],
                "summary": "Manually trigger update of current prices for holdings",
                "operationId": "bc93310022e828a88c5f1c264f56f5f2",
                "responses": {
                    "200": {
                        "description": "Price update job dispatched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Price update job dispatched successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClientHolding": {
                "required": [
                    "client_id",
                    "stock_symbol",
                    "quantity",
                    "purchase_price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "client_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "stock_symbol": {
                        "type": "string",
                        "example": "DSEXYZ"
                    },
                    "sector": {
                        "type": "string",
                        "example": "Dhaka"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 100
                    },
                    "purchase_price": {
                        "type": "number",
                        "format": "float",
                        "example": 75.5
                    },
                    "current_price": {
                        "type": "number",
                        "format": "float",
                        "example": 50.5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}